api,bus: address Chris review remarks

+++ bus/bus.go
@@ -433,6 +433,10 @@ func (b *bus) hostsRemoveHandlerPOST(jc jape.Context) {
 	if jc.Decode(&hrr) != nil {
 		return
 	}
+	if hrr.MaxDowntimeHours == 0 {
+		jc.Error(errors.New("maxDowntime must be non-zero"), http.StatusBadRequest)
+		return
+	}
 	removed, err := b.hdb.RemoveOfflineHosts(jc.Request.Context(), hrr.MinRecentScanFailures, time.Duration(hrr.MaxDowntimeHours))
 	if jc.Check("couldn't remove offline hosts", err) != nil {
 		return
